==> 10_model.Rmd <==
---
title: "Modeling"
author: "Rochelle Terman"
date: "December 14, 2015"
output: html_document
---

#### Table of Contents

1. Prepare Data
2. Logit (Hypothesis 1)
3. Negative Binomial (Hypothesis 1)
4. Heckit (Hypothesis 2)


```{r}
rm(list=ls())
setwd("~/Dropbox/berkeley/Git-Repos/worlds-women")

require(plyr)
require("MASS")
require("xtable")
require(plm) # panel data operations
require(ggplot2) # visualizations
require(stargazer) # pretty tables
require(sandwich) # for robust standard errors
require(lmtest) # for robust standard errors
require(sampleSelection) # for heckman correction
require(mfx) # for marginal effects
require(Matching) # required by the interaction_plots.R script below
source("interaction_plots.R") # for interaction plots; see paper for attribution
```

# 1. Prepare Data

```{r}
rt <- read.csv("Data/country-year/original.csv")

# subset
rt <- rt[,c("ccode", "year", "n.docs", "n.words", "n.binary", "rights", "alt.dv.perc", "muslim", "muslim.maj", "mena", "polity2", "physint", "amnesty", "statedept", "gdp.pc.un", "pop.wdi", "wopol", "wosoc", "wecon", "domestic9", "count", "region")]

# Deal with Palestine (without ccode)
rt$ccode[is.na(rt$ccode)] <- 999

# composite women's rights
rt$women_composite <- rowMeans(cbind(rt$wopol,rt$wosoc,rt$wecon), na.rm = T)

# Summarize Data
stargazer(rt, type="text", out = "Data/country-year/summary.txt")

# subset with an observation
rt.1 <- rt[rt$n.binary==1,]
summary(rt.1$muslim)

# israel
rt$mena[rt$ccode == 666] <- 0

# n.obs
nrow(rt[rt$year>1979,]) # 6226

# make panel data
names(rt)
rt <- pdata.frame(rt, c("ccode","year"))

# set lagged variables
rt$n.docs.lag <- lag(rt$n.docs, 1)
rt$n.binary.lag <- lag(rt$n.binary, 1)
rt$rights.lag <- lag(rt$rights, 1)
rt$alt.dv.perc.lag <- lag(rt$alt.dv.perc, 1)
rt$muslim.maj <- lag(rt$muslim.maj, 1)
rt$muslim <- lag(rt$muslim, 1)
rt$women_composite <- lag(rt$women_composite, 1)
rt$wopol <- lag(rt$wopol, 1)
rt$wosoc <- lag(rt$wosoc, 1)
rt$wecon <- lag(rt$wecon, 1)
rt$count <- lag(rt$count, 1)
rt$polity2 <- lag(rt$polity2, 1)
rt$domestic9 <- lag(rt$domestic9, 1)
rt$pop.wdi <- lag(rt$pop.wdi, 1)
rt$gdp.pc.un <- lag(rt$gdp.pc.un, 1)

# before, after 9/11
rt.before <- rt[as.numeric(levels(rt$year))<2002,]
rt.after <- rt[as.numeric(levels(rt$year))>=2002,]

# muslim maj and non-Muslim maj
muslim.maj <- rt[rt$muslim.maj == 1,]
muslim.maj <- muslim.maj[!is.na(muslim.maj$region)]
non.muslim.maj <- rt[rt$muslim.maj == 0,]
non.muslim.maj <- non.muslim.maj[!is.na(non.muslim.maj$region)]
```

# 2. Logit Model on Reports (Binary)

## 2.1 Full Models

```{r}
# MUSLIM MAJORITY
logit1 <- glm(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un), data = rt, family = binomial(link="probit"), control = list(maxit = 50))
summary(logit1)
coeftest(logit1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))
logit1.se <- coeftest(logit1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

jpeg("Results/regressions/interactive-plots/logit1.jpeg",width=700,height=500,type="quartz")
interaction_plot_binary(logit1, effect="women_composite", moderator="muslim.maj", interaction="women_composite:muslim.maj", factor_labels=c("Not Muslim","Muslim"), xlabel="Muslim majority", ylabel="Marginal Effect of Women's Rights on Coverage", title="")
dev.off()

# MENA
logit2 <- glm(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un), data = rt, family = binomial(link="probit"), control = list(maxit = 50))
logit2.se <- coeftest(logit2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

jpeg("Results/regressions/interactive-plots/logit2.jpeg",width=700,height=500,type="quartz")
interaction_plot_binary(logit2, effect="women_composite", moderator="mena", interaction="women_composite:mena", factor_labels=c("Not Mena","Mena"), xlabel="Mena", ylabel="Marginal Effect of Women's Rights on Coverage", title="")
dev.off()

# PERCENT MUSLIM
logit3 <- glm(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un), data = rt, family = binomial(link="probit"), control = list(maxit = 50))
logit3.se <- coeftest(logit3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

jpeg("Results/regressions/interactive-plots/logit3.jpeg",width=700,height=500,type="quartz")
interaction_plot_continuous(logit3, effect="women_composite", moderator="muslim", interaction="women_composite:muslim", mean=T, title="",xlabel="Percentage Muslim", ylabel="Marginal Effect of Women's Rights on Coverage")
dev.off()

# PRINT
stargazer(logit1, logit2, logit3, out = "Results/regressions/main-results/probit.html", type = "html", label = "table:probit", style = "ajps", title = "Probit Analysis of U.S. News Coverage about Women Abroad", se = list(logit1.se, logit2.se, logit3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, .001))

# muslim maj
probitmfx(n.binary ~ count + women_composite + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = muslim.maj,robust = T, atmean = F)

# non-muslim-maj
negbinmfx(n.binary ~ count + women_composite + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = non.muslim.maj,robust = T, atmean = F)
```

## 2.2 Partial Models

To respond to critiques that the findings are model dependent, this adds covariates one by one. Uses only Muslim majority for now. 

TODO: MENA and Muslim Percentage in Robustness tests.

```{r}
# basic model (women's rights * muslim majority)
part.logit.1 <- glm(n.binary ~ women_composite*muslim.maj, data = rt, family = binomial(link="probit"))
part.logit.1.se <- coeftest(part.logit.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add Country Reports
part.logit.2 <- glm(n.binary ~ women_composite*muslim.maj + count, data = rt, family = binomial(link="probit"))
part.logit.2.se <- coeftest(part.logit.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add Democracy
part.logit.3 <- glm(n.binary ~  women_composite*muslim.maj + count + polity2, data = rt, family = binomial(link="probit"))
part.logit.3.se <- coeftest(part.logit.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add Instability
part.logit.4 <- glm(n.binary ~  women_composite*muslim.maj + count + polity2 + domestic9, data = rt, family = binomial(link="probit"))
part.logit.4.se <- coeftest(part.logit.4, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add Population
part.logit.5 <- glm(n.binary ~  women_composite*muslim.maj + count + polity2 + domestic9 + log(pop.wdi), data = rt, family = binomial(link="probit"))
part.logit.5.se <- coeftest(part.logit.5, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add GDP
part.logit.6 <- glm(n.binary ~ women_composite*muslim.maj + count + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un), data = rt, family = binomial(link="probit"))
part.logit.6.se <- coeftest(part.logit.6, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# write
stargazer(part.logit.1, part.logit.2, part.logit.3, part.logit.4, part.logit.5, part.logit.6, out = "Results/regressions/robust/probit-part.html", type = "html", label = "table:probit-part", style = "ajps", title = "Probit Analysis of U.S. News Coverage about Women Abroad", se = list(part.logit.1.se, part.logit.2.se, part.logit.3.se, part.logit.4.se, part.logit.5.se, part.logit.6.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Women's Rights Index","Muslim Majority", "Country Reports", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Maj"), star.cutoffs = c(0.05, 0.01, .001))
```

## 2.3 OLS 

```{r}
lm.binary<- lm(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
summary(lm.binary)
coeftest(lm.binary, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))
```

## 2.4 Robustness Tests

### 2.4.1 Replacing Women's Rights Composit with Women's Political Rights

```{r}
# muslim majority
logit.r1.1 <- glm(n.binary ~ count + wopol*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r1.1.se <- coeftest(logit.r1.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
logit.r1.2 <- glm(n.binary ~ count + wopol*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r1.2.se <- coeftest(logit.r1.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
logit.r1.3 <- glm(n.binary ~ count + wopol*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r1.3.se <- coeftest(logit.r1.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(logit.r1.1, logit.r1.2, logit.r1.3, type = "html", out = "Results/regressions/robust/probit-1.html", style = "ajps", title = "H1.A: Women's Political Rights", se = list(logit.r1.1.se, logit.r1.2.se, logit.r1.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Political Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, .001))
```

### 2.4.2 Replacing Women's Rights Composit with Women's Social Rights

```{r}
# muslim majority
logit.r2.1 <- glm(n.binary ~ count + wosoc*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r2.1.se <- coeftest(logit.r2.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
logit.r2.2 <- glm(n.binary ~ count + wosoc*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r2.2.se <- coeftest(logit.r2.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
logit.r2.3 <- glm(n.binary ~ count + wosoc*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r2.3.se <- coeftest(logit.r2.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(logit.r2.1, logit.r2.2, logit.r2.3, type = "html", out = "Results/regressions/robust/probit-2.html", style = "ajps", title = "H1.A: Women's Social Rights", se = list(logit.r2.1.se, logit.r2.2.se, logit.r2.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Social Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, .001))
```

### 2.4.3 Replacing Women's Rights Composit with Women's Economic Rights

```{r}
# muslim majority
logit.r3.1 <- glm(n.binary ~ count + wecon*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r3.1.se <- coeftest(logit.r3.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
logit.r3.2 <- glm(n.binary ~ count + wecon*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r3.2.se <- coeftest(logit.r3.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
logit.r3.3 <- glm(n.binary ~ count + wecon*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt, family = "binomial")
logit.r3.3.se <- coeftest(logit.r3.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(logit.r3.1, logit.r3.2, logit.r3.3, type = "html", style = "ajps", out = "Results/regressions/robust/probit-3.html", title = "H1.A: Women's Economic Rights", se = list(logit.r3.1.se, logit.r3.2.se, logit.r3.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Economic Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, .001))
```

### 2.4.4 Adding Israel to MENA

```{r}
rt.israel <- rt
rt.israel$mena[rt.israel$ccode == 666] <- 1

# muslim majority
logit.r4.1 <- glm(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.israel, family = "binomial")
logit.r4.1.se <- coeftest(logit.r4.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
logit.r4.2 <- glm(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.israel, family = "binomial")
logit.r4.2.se <- coeftest(logit.r4.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
logit.r4.3 <- glm(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.israel, family = "binomial")
logit.r4.3.se <- coeftest(logit.r4.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(logit.r4.1, logit.r4.2, logit.r4.3, type = "html", out = "Results/regressions/robust/probit-4.html", style = "ajps", title = "H1.A: Israel Added to MENA", se = list(logit.r4.1.se, logit.r4.2.se, logit.r4.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, .001))
```

### 2.4.5 Adding Lagged DV

```{r}
# MUSLIM MAJORITY
logit.r5.1 <- glm(n.binary ~ n.binary.lag + count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un), data = rt, family = binomial(link="probit"))
logit.r5.1.se <- coeftest(logit.r5.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# MENA
logit.r5.2 <- glm(n.binary ~ n.binary.lag + count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un), data = rt, family = binomial(link="probit"))
logit.r5.2.se <- coeftest(logit.r5.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PERCENT MUSLIM
logit.r5.3 <- glm(n.binary ~ n.binary.lag + count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un), data = rt, family = binomial(link="probit"), control = list(maxit = 50))
logit.r5.3.se <- coeftest(logit.r5.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(logit.r5.1, logit.r5.2, logit.r5.3, type = "html", style = "ajps", out = "Results/regressions/robust/probit-5.html", label = "table:probit", title = "H1.A: Lagged DV", se = list(logit.r5.1.se, logit.r5.2.se, logit.r5.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Lagged DV", "Country Reports", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, .001))
```

### 2.4.6 Before  9/11 

```{r}
# muslim majority
logit.r6.1 <- glm(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.before, family = "binomial")
logit.r6.1.se <- coeftest(logit.r6.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
logit.r6.2 <- glm(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.before, family = "binomial")
logit.r6.2.se <- coeftest(logit.r6.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
logit.r6.3 <- glm(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.before, family = "binomial")
logit.r6.3.se <- coeftest(logit.r6.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(logit.r6.1, logit.r6.2, logit.r6.3, type = "html", out = "Results/regressions/robust/probit-6.html", title = "H1.A: Before 911", style = "ajps", se = list(logit.r6.1.se, logit.r6.2.se, logit.r6.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001))

# VISUALIZE
interaction_plot_binary(logit.r6.1, effect="women_composite", moderator="muslim.maj", interaction="women_composite:muslim.maj", factor_labels=c("Not Muslim","Muslim"), xlabel="Muslim majority", ylabel="Marginal Effect of Women's Rights on Coverage", title="")

interaction_plot_binary(logit.r6.2, effect="women_composite", moderator="mena", interaction="women_composite:mena", factor_labels=c("Not Mena","Mena"), xlabel="Mena", ylabel="Marginal Effect of Women's Rights on Coverage", title="")

interaction_plot_continuous(logit.r6.3, effect="women_composite", moderator="muslim", interaction="women_composite:muslim", mean=T, title="",xlabel="Percentage Muslim", ylabel="Marginal Effect of Women's Rights on Coverage")
```

### 2.4.7 After  9/11 

```{r}
# muslim majority
logit.r7.1 <- glm(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.after, family = "binomial")
logit.r7.1.se <- coeftest(logit.r7.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
logit.r7.2 <- glm(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.after, family = "binomial")
logit.r7.2.se <- coeftest(logit.r7.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
logit.r7.3 <- glm(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.after, family = "binomial")
logit.r7.3.se <- coeftest(logit.r7.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(logit.r7.1, logit.r7.2, logit.r7.3, type = "html", out = "Results/regressions/robust/probit-7.html", title = "H1.A: After 911", style = "ajps", se = list(logit.r7.1.se, logit.r7.2.se, logit.r7.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001))

# VISUALIZE
interaction_plot_binary(logit.r7.1, effect="women_composite", moderator="muslim.maj", interaction="women_composite:muslim.maj", factor_labels=c("Not Muslim","Muslim"), xlabel="Muslim majority", ylabel="Marginal Effect of Women's Rights on Coverage", title="")

interaction_plot_binary(logit.r7.2, effect="women_composite", moderator="mena", interaction="women_composite:mena", factor_labels=c("Not Mena","Mena"), xlabel="Mena", ylabel="Marginal Effect of Women's Rights on Coverage", title="")

interaction_plot_continuous(logit.r7.3, effect="women_composite", moderator="muslim", interaction="women_composite:muslim", mean=T, title="",xlabel="Percentage Muslim", ylabel="Marginal Effect of Women's Rights on Coverage")
```

## 2.4.8 Year Fixed Effects

```{r}
# MUSLIM MAJORITY
logit.r8.1 <- glm(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un) + as.factor(year), data = rt, family = binomial(link="probit"), control = list(maxit = 50))
summary(logit.r8.1)
coeftest(logit.r8.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))
logit.r8.1.se <- coeftest(logit.r8.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# MENA
logit.r8.2 <- glm(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un) + as.factor(year), data = rt, family = binomial(link="probit"), control = list(maxit = 50))
logit.r8.2.se <- coeftest(logit.r8.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PERCENT MUSLIM
logit.r8.3 <- glm(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) + log(gdp.pc.un) + as.factor(year), data = rt, family = binomial(link="probit"), control = list(maxit = 50))
logit.r8.3.se <- coeftest(logit.r8.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]
```

# 3. Negative Binomial Model on Reports (Count)

## 3.1 Full Models

```{r}
# MUSLIM MAJORITY
nb1 <- glm.nb(n.docs ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
summary(nb1)
nb1.se <- coeftest(nb1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

pdf("Results/regressions/interactive-plots/nb1.pdf",width=7,height=5)
interaction_plot_binary(nb1, effect="women_composite", moderator="muslim.maj", interaction="women_composite:muslim.maj", factor_labels=c("Not Muslim Majority","Muslim Majority"), xlabel="", ylabel="Marginal Effect of Women's Rights on Coverage", title="")
dev.off()

# MENA
nb2 <- glm.nb(n.docs ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
summary(nb2)
nb2.se <- coeftest(nb2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

pdf("Results/regressions/interactive-plots/nb2.pdf",width=7,height=5)
interaction_plot_binary(nb2, effect="women_composite", moderator="mena", interaction="women_composite:mena", factor_labels=c("Not Mena","Mena"), xlabel="", ylabel="Marginal Effect of Women's Rights on Coverage", title="")
dev.off()

# MUSLIM PERCENTAGE
nb3 <- glm.nb(n.docs ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
summary(nb3)
nb3.se <- coeftest(nb3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

pdf("Results/regressions/interactive-plots/nb3.pdf",width=7,height=5)
interaction_plot_continuous(nb3, effect="women_composite", moderator="muslim", interaction="women_composite:muslim", mean=T, title="",xlabel="Percentage Muslim", ylabel="Marginal Effect of Women's Rights on Coverage")
dev.off()

# PRINT
stargazer(nb1, nb2, nb3, type = "html", out = "Results/regressions/main-results/negbin.html", label = "table:negbin", title = "Negative Binomial Analysis of U.S. News Coverage about Women Abroad", style = "ajps", se = list(nb1.se, nb2.se, nb3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Count)", covariate.labels=c("Country Reports", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001))

# interpret for Muslim-maj and non-Muslim-maj obs

# muslim maj
negbinmfx(n.docs ~ count + women_composite + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = muslim.maj,robust = T, atmean = F)

# non-muslim-maj
negbinmfx(n.docs ~ count + women_composite + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = non.muslim.maj,robust = T, atmean = F)
```

## 3.2 Partial Models

To respond to critiques that the findings are model dependent, this adds covariates one by one. Uses only Muslim majority for now. 

TODO: MENA and Muslim Percentage in Robustness tests.

```{r}
# basic model (women's rights * muslim majority)
part.nb.1 <- glm.nb(n.docs ~ women_composite*muslim.maj, data = rt)
part.nb.1.se <- coeftest(part.nb.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add Country Reports
part.nb.2 <- glm.nb(n.docs ~ women_composite*muslim.maj + count, data = rt)
part.nb.2.se <- coeftest(part.nb.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add Democracy
part.nb.3 <- glm.nb(n.docs ~ women_composite*muslim.maj + count + polity2, data = rt)
part.nb.3.se <- coeftest(part.nb.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add Instability
part.nb.4 <- glm.nb(n.docs ~ women_composite*muslim.maj + count + polity2 + lag(domestic9, 1), data = rt)
part.nb.4.se <- coeftest(part.nb.4, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add Population
part.nb.5 <- glm.nb(n.docs ~ women_composite*muslim.maj + count + polity2 + lag(domestic9, 1) + log(pop.wdi), data = rt)
part.nb.5.se <- coeftest(part.nb.5, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# add GDP
part.nb.6 <- glm.nb(n.docs ~ women_composite*muslim.maj + count + polity2 + lag(domestic9, 1) + log(pop.wdi) +log(gdp.pc.un), data = rt)
part.nb.6.se <- coeftest(part.nb.6, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# write
stargazer(part.nb.1, part.nb.2, part.nb.3, part.nb.4, part.nb.5, part.nb.6, out = "Results/regressions/robust/negbin-part.html", type = "html", label = "table:negbin-part", style = "ajps", title = "Probit Analysis of U.S. News Coverage about Women Abroad", se = list(part.nb.1.se, part.nb.2.se, part.nb.3.se, part.nb.4.se, part.nb.5.se, part.nb.6.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Count)", covariate.labels=c("Women's Rights Index","Muslim Majority", "Country Reports", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Maj"), star.cutoffs = c(0.05, 0.01, .001))
```

## 3.3 OLS

```{r}
# with lagged DV
lm.count.1 <- lm(n.docs ~ n.docs.lag + count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
summary(lm.count.1)
coeftest(lm.count.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))

# without lagged DV
lm.count.2 <- lm(n.docs ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
summary(lm.count.2)
coeftest(lm.count.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))
```

## 3.4 Robustness Tests

### 3.4.1 Replacing Women's Rights Composit with Women's Political Rights

```{r}
# muslim majority
nb.r1.1 <- glm.nb(n.docs ~ count + wopol*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
summary(nb.r1.1)
nb.r1.1.se <- coeftest(nb.r1.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
nb.r1.2 <- glm.nb(n.docs ~ count + wopol*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r1.2.se <- coeftest(nb.r1.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
nb.r1.3 <- glm.nb(n.docs ~ count + wopol*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r1.3.se <- coeftest(nb.r1.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(nb.r1.1, nb.r1.2, nb.r1.3, type = "html",  out = "Results/regressions/robust/negbin-1.html", title = "H1.B: Women's Political Rights", style = "ajps", se = list(nb.r1.1.se, nb.r1.2.se, nb.r1.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Count)", covariate.labels=c("Country Reports", "Women's Political Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001), omit.stat = "theta")
```

### 3.4.2 Replacing Women's Rights Composit with Women's Social Rights

```{r}
# muslim majority
nb.r2.1 <- glm.nb(n.docs ~ count + wosoc*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r2.1.se <- coeftest(nb.r2.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
nb.r2.2 <- glm.nb(n.docs ~ count + wosoc*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r2.2.se <- coeftest(nb.r2.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
nb.r2.3 <- glm.nb(n.docs ~ count + wosoc*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r2.3.se <- coeftest(nb.r2.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(nb.r2.1, nb.r2.2, nb.r2.3, type = "html", out = "Results/regressions/robust/negbin-2.html", title = "H1.B: Women's Social Rights", style = "ajps", se = list(nb.r2.1.se, nb.r2.2.se, nb.r2.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Count)", covariate.labels=c("Country Reports", "Women's Social Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001), omit.stat = "theta")
```

### 3.4.3 Replacing Women's Rights Composit with Women's Economic Rights

```{r}
# muslim majority
nb.r3.1 <- glm.nb(n.docs ~ count + wecon*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r3.1.se <- coeftest(nb.r3.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
nb.r3.2 <- glm.nb(n.docs ~ count + wecon*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r3.2.se <- coeftest(nb.r3.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
nb.r3.3 <- glm.nb(n.docs ~ count + wecon*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r3.3.se <- coeftest(nb.r3.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(nb.r3.1, nb.r3.2, nb.r3.3, type = "html", out = "Results/regressions/robust/negbin-3.html", title = "H1.B: Women's Economic Rights", style = "ajps", se = list(nb.r3.1.se, nb.r3.2.se, nb.r3.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Count)", covariate.labels=c("Country Reports", "Women's Economic Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001), omit.stat = "theta")
```

### 3.4.4 Adding Israel to MENA

```{r}
rt.israel <- rt
rt.israel$mena[rt.israel$ccode == 666] <- 1

# muslim majority
nb.r4.1 <- glm.nb(n.docs ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.israel)
nb.r4.1.se <- coeftest(nb.r4.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
nb.r4.2 <- glm.nb(n.docs ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.israel)
nb.r4.2.se <- coeftest(nb.r4.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
nb.r4.3 <- glm.nb(n.docs ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.israel)
nb.r4.3.se <- coeftest(nb.r4.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(nb.r4.1, nb.r4.2, nb.r4.3, type = "html", out = "Results/regressions/robust/negbin-4.html", title = "H1.B: Adding Israel to MENA", style = "ajps", se = list(nb.r4.1.se, nb.r4.2.se, nb.r4.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Count)", covariate.labels=c("Country Reports", "Women's Economic Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001), omit.stat = "theta")
```

### 3.4.5 Adding Lagged DV

```{r}
# muslim majority
nb.r5.1 <- glm.nb(n.docs ~ n.docs.lag + count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r5.1.se <- coeftest(nb.r5.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
nb.r5.2 <- glm.nb(n.docs ~ n.docs.lag + count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r5.2.se <- coeftest(nb.r5.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
nb.r5.3 <- glm.nb(n.docs ~ n.docs.lag + count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt)
nb.r5.3.se <- coeftest(nb.r5.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(nb.r5.1, nb.r5.2, nb.r5.3, type = "html", out = "Results/regressions/robust/negbin-5.html", title = "H1.B: Lagged DV", style = "ajps", se = list(nb.r5.1.se, nb.r5.2.se, nb.r5.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Count)", covariate.labels=c("Lagged DV", "Country Reports", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001), omit.stat = "theta")
```

### 3.4.6 Before  9/11 

```{r}
rt.before <- rt[as.numeric(levels(rt$year))<2002,]
rt.after <- rt[as.numeric(levels(rt$year))>=2002,]

# muslim majority
nb.r6.1 <- glm.nb(n.docs ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un), data = rt.before)
nb.r6.1.se <- coeftest(nb.r6.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
nb.r6.2 <- glm.nb(n.docs ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un), data = rt.before)
nb.r6.2.se <- coeftest(nb.r6.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
nb.r6.3 <- glm.nb(n.docs ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un), data = rt.before)
nb.r6.3.se <- coeftest(nb.r6.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(nb.r6.1, nb.r6.2, nb.r6.3, type = "html", out = "Results/regressions/robust/negbin-6.html", title = "H2.A: Before 911", style = "ajps", se = list(nb.r6.1.se, nb.r6.2.se, nb.r6.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001))
```

### 3.4.7 After  9/11 

```{r}
rt.after <- rt[as.numeric(levels(rt$year))>=2002,]

# muslim majority
nb.r7.1 <- glm.nb(n.docs ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.after)
nb.r7.1.se <- coeftest(nb.r7.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
nb.r7.2 <- glm.nb(n.docs ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.after)
nb.r7.2.se <- coeftest(nb.r7.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim percentage
nb.r7.3 <- glm.nb(n.docs ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),data = rt.after)
nb.r7.3.se <- coeftest(nb.r7.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(nb.r7.1, nb.r7.2, nb.r7.3, type = "html", out = "Results/regressions/robust/negbin-7.html", title = "H1.A: After 911", style = "ajps", se = list(nb.r7.1.se, nb.r7.2.se, nb.r7.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Reported (Binary)", covariate.labels=c("Country Reports", "Women's Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Instability", "Population", "GDP per capita", "Women's Rights x Muslim Majority", "Women's Rights x MENA","Women's Rights x Muslim Percentage"), star.cutoffs = c(0.05, 0.01, 0.001))
```

# 4. Heckit Models

## 4.1 Full Models

```{r}
## Muslim Majority
heckit1 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                 rights ~  women_composite + muslim.maj + polity2 + physint,
                 rt )
summary(heckit1)
summary(heckit1$lm)
heckit1.se = coeftest(heckit1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# Coefficient Plot
require(coefplot)
pdf(file = "Results/regressions/main-results/heckit1-coefplot.pdf", width = 8, height = 5)
coefplot(heckit1$lm, title = "", coefficients = c("XOwomen_composite", "XOmuslim.maj", "XOpolity2", "XOphysint"), newNames = c(XOwomen_composite = "Women's Rights Index", XOmuslim.maj = "Muslim Majority", XOpolity2 = "Physical Integrity Rights", XOphysint = "Democracy"))
dev.off()

# MENA
heckit2 <- heckit(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + mena + polity2 + physint,
                  rt )
heckit2.se = coeftest(heckit2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# Muslim percentage
heckit3 <- heckit(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + muslim + polity2 + physint,
                  rt )
summary(heckit3$lm)
heckit3.se = coeftest(heckit3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]


# print - HTML
stargazer(heckit1$lm, heckit2$lm, heckit3$lm, type = "html", out = "Results/regressions/main-results/heckit.html", label = "table:heckit", style = "ajps", title = "Two-Step Analysis of Rights Focus in U.S. News Coverage about Women Abroad", se = list(heckit1.se, heckit2.se, heckit3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights Focus", covariate.labels=c("Intercept", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001))
```

## 4.2 Partial Models

```{r}
# Basic model - rights and muslim.maj
part.heckit.1 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                 rights ~ women_composite + muslim.maj,
                 rt )
part.heckit.1.se = coeftest(part.heckit.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# Add Democracy
part.heckit.2 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                 rights ~ women_composite + muslim.maj + polity2,
                 rt )
part.heckit.2.se = coeftest(part.heckit.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# Add Physical Integrity
part.heckit.3 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                 rights ~ women_composite + muslim.maj + polity2 + physint,
                 rt )
part.heckit.3.se = coeftest(part.heckit.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# print
stargazer(part.heckit.1$lm, part.heckit.2$lm, part.heckit.3$lm, type = "html", out = "Results/regressions/robust/heckit-part.html", style = "ajps", title = "Two-Step Analysis of Rights Focus in U.S. News Coverage about Women Abroad", se = list(part.heckit.1.se, part.heckit.2.se, part.heckit.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights Focus", covariate.labels=c("Intercept", "Women's Rights Index","Muslim Majority", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001))
```

## 4.2 OLS

```{r}
# with lagged DV
ols.rights <- lm(rights ~ women_composite + muslim.maj + polity2 + physint, data = rt) 
summary(ols.rights)
coeftest(ols.rights, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))

# without lagged DV
ols.rights <- lm(rights ~ women_composite + muslim.maj + polity2 + physint, data = rt) 
summary(ols.rights)
coeftest(ols.rights, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))
```

## 4.4 Robustness Tests

### 4.4.1 Replacing Women's Rights Composit with Women's Political Rights

```{r}
# muslim maj
heckit.r1.1 <- heckit(n.binary ~ count + wopol*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wopol + muslim.maj + polity2 + physint,
                      rt)
heckit.r1.1.se <- coeftest(heckit.r1.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
heckit.r1.2 <- heckit(n.binary ~ count + wopol*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wopol + mena + polity2 + physint,
                      rt )
heckit.r1.2.se <- coeftest(heckit.r1.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim perc.
heckit.r1.3 <- heckit(n.binary ~ count + wopol*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wopol + muslim + polity2 + physint,
                      rt )
heckit.r1.3.se <- coeftest(heckit.r1.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT 
stargazer(heckit.r1.1$lm, heckit.r1.2$lm, heckit.r1.3$lm, type = "html", out = "Results/regressions/robust/heckit-1.html", style = "ajps", title = "H2: Women's Political Rights", se = list(heckit.r1.1.se, heckit.r1.2.se, heckit.r1.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights Focus", covariate.labels=c("Intercept", "Women's Political Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.2 Replacing Women's Rights Composit with Women's Social Rights

```{r}
# muslim maj
heckit.r2.1 <- heckit(n.binary ~ count + wosoc*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wosoc + muslim.maj + polity2 + physint,
                      rt)
heckit.r2.1.se <- coeftest(heckit.r2.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
heckit.r2.2 <- heckit(n.binary ~ count + wosoc*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wosoc + mena + polity2 + physint,
                      rt )
heckit.r2.2.se <- coeftest(heckit.r2.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim perc.
heckit.r2.3 <- heckit(n.binary ~ count + wosoc*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wosoc + muslim + polity2 + physint,
                      rt )
heckit.r2.3.se <- coeftest(heckit.r2.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(heckit.r2.1$lm, heckit.r2.2$lm, heckit.r2.3$lm, type = "html", out = "Results/regressions/robust/heckit-2.html", style = "ajps", title = "H2: Women's Social Rights", se = list(heckit.r2.1.se, heckit.r2.2.se, heckit.r2.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights Focus", covariate.labels=c("Intercept", "Women's Social Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.3 Replacing Women's Rights Composit with Women's Economic Rights

```{r}
# muslim maj
heckit.r3.1 <- heckit(n.binary ~ count + wecon*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wecon + muslim.maj + polity2 + physint,
                      rt)
heckit.r3.1.se <- coeftest(heckit.r3.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
heckit.r3.2 <- heckit(n.binary ~ count + wecon*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wecon + mena + polity2 + physint,
                      rt )
heckit.r3.2.se <- coeftest(heckit.r3.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim perc.
heckit.r3.3 <- heckit(n.binary ~ count + wecon*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ wecon + muslim + polity2 + physint,
                      rt )
heckit.r3.3.se <- coeftest(heckit.r3.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(heckit.r3.1$lm, heckit.r3.2$lm, heckit.r3.3$lm, type = "html", out = "Results/regressions/robust/heckit-3.html", style = "ajps", title = "H2: Women's Economic Rights", se = list(heckit.r3.1.se, heckit.r3.2.se, heckit.r3.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights Focus", covariate.labels=c("Intercept", "Women's Economic Rights","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.4 Adding Israel to MENA

```{r}
rt.israel <- rt
rt.israel$mena[rt.israel$ccode == 666] <- 1

# muslim maj
heckit.r4.1 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ women_composite + muslim.maj + polity2 + physint,
                      rt.israel)
heckit.r4.1.se <- coeftest(heckit.r4.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
heckit.r4.2 <- heckit(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ women_composite + mena + polity2 + physint,
                      rt.israel )
heckit.r4.2.se <- coeftest(heckit.r4.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim perc.
heckit.r4.3 <- heckit(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ women_composite + muslim + polity2 + physint,
                      rt.israel )
heckit.r4.3.se <- coeftest(heckit.r4.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(heckit.r4.1$lm, heckit.r4.2$lm, heckit.r4.3$lm, out = "Results/regressions/robust/heckit-4.html", type = "html", style = "ajps", title = "H2: Adding Israel to MENA", se = list(heckit.r4.1.se, heckit.r4.2.se, heckit.r4.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights Focus", covariate.labels=c("Intercept", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physican Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.5 Adding Lagged DV

```{r}
# muslim maj
heckit.r5.1 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ rights.lag + women_composite + muslim.maj + polity2 + physint,
                      rt)
heckit.r5.1.se <- coeftest(heckit.r5.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# mena
heckit.r5.2 <- heckit(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ rights.lag + women_composite + mena + polity2 + physint,
                      rt )
summary(heckit.r5.2$lm)
heckit.r5.2.se <- coeftest(heckit.r5.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# muslim perc.
heckit.r5.3 <- heckit(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      rights ~ rights.lag + women_composite + muslim + polity2 + physint,
                      rt )
heckit.r5.3.se <- coeftest(heckit.r5.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(heckit.r5.1$lm, heckit.r5.2$lm, heckit.r5.3$lm, out = "Results/regressions/robust/heckit-5.html", type = "html", style = "ajps", title = "H2: Lagged DV", se = list(heckit.r5.1.se, heckit.r5.2.se, heckit.r5.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Lagged DV", "Rights Focus", covariate.labels=c("Intercept", "Lagged DV", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physican Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.6 Fractional Logit 

```{r}
heckit.r6.1 <- glm(rights ~ women_composite + muslim.maj + polity2 + physint, data = rt, family=binomial(link="logit"), weights = n.words)
summary(heckit.r6.1)
heckit.r6.1.se <- coeftest(heckit.r6.1, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

heckit.r6.2 <- glm(rights ~ women_composite + mena + polity2 + physint, data = rt, family=binomial(link="logit"), weights = n.words)
summary(heckit.r6.2)
heckit.r6.2.se <- coeftest(heckit.r6.2, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

heckit.r6.3 <- glm(rights ~ women_composite + muslim + polity2 + physint, data = rt, family=binomial(link="logit"), weights = n.words)
summary(heckit.r6.3)
heckit.r6.3.se <- coeftest(heckit.r6.3, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

stargazer(heckit.r6.1, heckit.r6.2, heckit.r6.3, type = "html", out = "Results/regressions/robust/heckit-6.html", style = "ajps", title = "H2: Fractional Logit", se = list(heckit.r6.1.se, heckit.r6.2.se, heckit.r6.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights Focus", covariate.labels=c("Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.7 Use Alternative DV instead of Rights Means

```{r}
names(rt)

# Muslim Majority
heckit.r7.1 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      alt.dv.perc ~ women_composite + muslim.maj + polity2 + physint,
                      rt )
heckit.r7.1.se = coeftest(heckit.r7.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]
heckit.r7.1.se

# MENA
heckit.r7.2 <- heckit(n.binary ~ count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      alt.dv.perc ~  women_composite + mena + polity2 + physint,
                      rt )
heckit.r7.2.se = coeftest(heckit.r7.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# Muslim percentage
heckit.r7.3 <- heckit(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                      alt.dv.perc ~ women_composite + muslim + polity2 + physint,
                      rt )
summary(heckit.r7.3$lm)
heckit.r7.3.se = coeftest(heckit.r7.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(heckit.r7.1$lm, heckit.r7.2$lm, heckit.r7.3$lm, type = "html", out = "Results/regressions/robust/heckit-7.html", style = "ajps", title = "H2: With Alternative Estimate of DV", se = list(heckit.r7.1.se, heckit.r7.2.se, heckit.r7.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights (Binary)", covariate.labels=c("Intercept", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.8 Before 911

```{r}
# Muslim Majority
heckit.r8.1 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + muslim.maj + polity2 + physint,
                  rt.before )
summary(heckit.r8.1$lm)
heckit.r8.1.se = coeftest(heckit.r8.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]
heckit.r8.1.se

 # MENA
heckit.r8.2 <- heckit(n.binary ~count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + mena + polity2 + physint,
                  rt.before )
heckit.r8.2.se = coeftest(heckit.r8.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# Muslim percentage
heckit.r8.3 <- heckit(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + muslim + polity2 + physint,
                  rt.before )
summary(heckit.r8.3$lm)
heckit.r8.3.se = coeftest(heckit.r8.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(heckit.r8.1$lm, heckit.r8.2$lm, heckit.r8.3$lm, type = "html", out = "Results/regressions/robust/heckit-8.html", style = "ajps", title = "H2: Before 911", se = list(heckit.r8.1.se, heckit.r8.2.se, heckit.r8.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights (Binary)", covariate.labels=c("Intercept", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.9 After 911

```{r}
# Muslim Majority
heckit.r9.1 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + muslim.maj + polity2 + physint,
                  rt.after )
heckit.r9.1.se = coeftest(heckit.r9.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]
heckit.r9.1.se

 # MENA
heckit.r9.2 <- heckit(n.binary ~count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + mena + polity2 + physint,
                  rt.after )
heckit.r9.2.se = coeftest(heckit.r9.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# Muslim percentage
heckit.r9.3 <- heckit(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + muslim + polity2 + physint,
                  rt.after )
summary(heckit.r9.3$lm)
heckit.r9.3.se = coeftest(heckit.r9.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(heckit.r9.1$lm, heckit.r9.2$lm, heckit.r9.3$lm, type = "html", out = "Results/regressions/robust/heckit-9.html", style = "ajps", title = "H2: After 911", se = list(heckit.r9.1.se, heckit.r9.2.se, heckit.r9.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights (Binary)", covariate.labels=c("Intercept", "Women's Rights Index","Muslim Majority","MENA", "Muslim Percentage", "Democracy", "Physical Integrity Rights"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```

### 4.4.10 Interaction

```{r}
rt$after.911 <- 0
rt$after.911[as.numeric(levels(rt$year))>=2002] <- 1
rt$after.911 <- as.factor(rt$after.911)
length(which(rt$after.911==1))

# Muslim Majority
heckit.r10.1 <- heckit(n.binary ~ count + women_composite*muslim.maj + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + muslim.maj:after.911 + polity2 + physint,
                  rt )
heckit.r10.1.se = coeftest(heckit.r10.1$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]
heckit.r10.1.se

 # MENA
heckit.r10.2 <- heckit(n.binary ~count + women_composite*mena + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + mena:after.911 + polity2 + physint,
                  rt )
heckit.r10.2.se = coeftest(heckit.r10.2$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# Muslim percentage
heckit.r10.3 <- heckit(n.binary ~ count + women_composite*muslim + polity2 + domestic9 + log(pop.wdi) +log(gdp.pc.un),
                  rights ~ women_composite + muslim:after.911  + polity2 + physint,
                  rt )
summary(heckit.r10.3$lm)
heckit.r10.3.se = coeftest(heckit.r10.3$lm, vcov=function(x) vcovHC(x, cluster="group", type="HC1"))[,2]

# PRINT
stargazer(heckit.r10.1$lm, heckit.r10.2$lm, heckit.r10.3$lm, type = "html", out = "Results/regressions/robust/heckit-10.html", style = "ajps", title = "H2: Interactive 911", se = list(heckit.r10.1.se, heckit.r10.2.se, heckit.r10.3.se), notes="Robust standard errors clustered on country appear in parentheses.",  dep.var.labels = "Rights (Binary)", covariate.labels=c("Intercept", "Women's Rights Index", "Democracy", "Physical Integrity Rights", "Muslim Majority: Pre 9/11", "Muslim Majority: Post 9/11", "MENA: Pre 9/11", "MENA: Post 9/11", "Muslim Percentage: Post 9/11", "Muslim Percentage: Pre 9/11"), star.cutoffs = c(0.05, 0.01, 0.001), df = F)
```
